src/cli/pyproject.toml
[project]
name = "qtcloud-think"
version = "0.1.0"
description = "思维收集与澄清工具"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "typer>=0.12.0",
]

[tool.uv]
index-url = "https://mirrors.aliyun.com/pypi/simple"
